update typography stories
