--- packages/component-library/stories/styleGuideStories/TypographyStyleCommonSample.story.js
@@ -1,35 +1,0 @@
-import React from "react";
-import { storybookStyles } from "../storyStyles";
-
-const TypographyStyleCommonSample = () => (
-  <div style={storybookStyles.main}>
-    <h2>Common Sample:</h2>
-    <h1 className="Title">Headline (Rubik Light)</h1>
-    <p>Meta data (Merriweather)</p>
-    <h2>Sub heading (Rubik Bold)</h2>
-    <p>Paragraph (Merriweather)</p>
-    <h4>
-      <a href="https://www.youtube.com/watch?v=gJ5hbJsN7Hc">Action</a> (Rubik
-      Medium)
-    </h4>
-    <hr />
-    <h2>Example of a Common Sample:</h2>
-    <h1 className="Title">The Effect of the Minimum Wage throughout Oregon</h1>
-    <p>March 31st, 2017 | Published by Hack Oregon</p>
-    <h2>How does it affect me?</h2>
-    <p>
-      Vivamus vitae odio tempor neque iaculis iaculis. Vestibulum ante ipsum
-      primis in faucibus orci luctus et ultrices posuere cubilia Curae; Aliquam
-      gravida risus nibh, non rutrum ipsum efficitur at. Fusce volutpat nisl nec
-      odio aliquet facilisis. Integer eu lacus metus. Mauris tortor lectus,
-      pulvinar a aliquet vel, venenatis a diam. Vivamus et lorem feugiat,
-      venenatis justo vel, luctus eros.
-    </p>
-    <h4>
-      <a href="https://www.youtube.com/watch?v=gJ5hbJsN7Hc">Read More</a>
-    </h4>
-    <hr />
-  </div>
-);
-
-export default TypographyStyleCommonSample;
