--- packages/component-library/stories/styleGuideStories/TypographyStyleHeadings.story.js
@@ -1,17 +1,0 @@
-import React from "react";
-import { storybookStyles } from "../storyStyles";
-
-const TypographyStyleHeadings = () => (
-  <div style={storybookStyles.main}>
-    <h2>Headings:</h2>
-    <h1>Heading 1 Rubik Light in 36 point</h1>
-    <h2>Heading 2 Rubik Bold in 21 point</h2>
-    <h3>Heading 3 Rubik Black in 18 point</h3>
-    <h4>Heading 4 Rubik Black in 16 point</h4>
-    <h5>Heading 5 Rubik Bold in 13 point</h5>
-    <h6>Heading 6 Rubik Bold in 11 point</h6>
-    <hr />
-  </div>
-);
-
-export default TypographyStyleHeadings;
